add GenRepeatedBlobName
