+++ oneflow/core/operator/boxing_op.cpp
@@ -16,14 +16,14 @@ void BoxingOp::InitFromOpConf() {
   const BoxingOpConf& boxing_conf = op_conf().boxing_conf();
 
   for (int32_t i = 0; i < boxing_conf.in_num(); ++i) {
-    EnrollInputBn("in_" + std::to_string(i), false);
+    EnrollInputBn(GenRepeatedBlobName("in", i), false);
   }
   if (boxing_conf.in_box_case() == BoxingOpConf::kAddBox
       && boxing_conf.out_box_case() == BoxingOpConf::kSplitBox) {
     EnrollDataTmpBn("middle");
   }
   for (int32_t i = 0; i < boxing_conf.out_num(); ++i) {
-    EnrollOutputBn("out_" + std::to_string(i), false);
+    EnrollOutputBn(GenRepeatedBlobName("out", i), false);
   }
 }
