+++ oneflow/core/operator/normal_model_update_op.cpp
@@ -6,7 +6,7 @@ namespace oneflow {
 
 void NormalModelUpdtOp::InitFromOpConf() {
   FOR_RANGE(int32_t, i, 0, op_conf().normal_mdupdt_conf().in_num()) {
-    EnrollInputBn("in_" + std::to_string(i), false);
+    EnrollInputBn(GenRepeatedBlobName("in", i), false);
   }
   EnrollOutputBn("model", false);
   MdUpdtVirtualInitFromOpConf();
