+++ oneflow/core/operator/reduce_gather_op.cpp
@@ -6,7 +6,7 @@ namespace oneflow {
 void ReduceGatherOp::InitFromOpConf() {
   CHECK(op_conf().has_reduce_gather_conf());
   for (int32_t i = 0; i < op_conf().reduce_gather_conf().in_num(); ++i) {
-    EnrollInputBn("in_" + std::to_string(i), false);
+    EnrollInputBn(GenRepeatedBlobName("in", i), false);
   }
   EnrollOutputBn("out", false);
 }
